pyproject.toml
[project]
name = "telegram-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "python-telegram-bot>=22.3",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
